+++ packages/orm/angel_orm_generator/lib/src/migration_generator.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'package:analyzer/dart/constant/value.dart';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:angel3_model/angel3_model.dart';
@@ -198,7 +199,20 @@ class MigrationGenerator extends GeneratorForAnnotation<Orm> {
 
               var defaultValue = ctx.buildContext.defaults[name];
 
-              if (defaultValue != null && !defaultValue.isNull) {
+              // Handle 'defaultValue' on Column annotation
+              if (col.defaultValue != null) {
+                var defaultCode =
+                    dartObjectToString(col.defaultValue as DartObject);
+
+                if (defaultCode != null) {
+                  Expression defaultExpr = CodeExpression(
+                    Code(defaultCode),
+                  );
+                  cascade.add(refer('defaultsTo').call([defaultExpr]));
+                }
+
+                // Handle 'defaultValue' on SerializableField annotation
+              } else if (defaultValue != null && !defaultValue.isNull) {
                 var type = defaultValue.type;
                 Expression? defaultExpr;
 
@@ -219,9 +233,12 @@ class MigrationGenerator extends GeneratorForAnnotation<Orm> {
                     // Definitely an analyzer issue.
                   }
                 } else {
-                  defaultExpr = CodeExpression(
-                    Code(dartObjectToString(defaultValue)!),
-                  );
+                  var defaultCode = dartObjectToString(defaultValue);
+                  if (defaultCode != null) {
+                    defaultExpr = CodeExpression(
+                      Code(defaultCode),
+                    );
+                  }
                 }
 
                 if (defaultExpr != null) {